# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /user:
    # binds a127 app logic to a route
    x-swagger-router-controller: userController
    
    
    get:
      description: Returns all the users
      # used as the method name of the controller
      operationId: getUsers
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
    post:
      description: Add new User
      # used as the method name of the controller
      operationId: addUser
      parameters:
        - in: body
          name: user
          description: The User to create
          schema:
            type: object
            required:
              - firstName
              - lastName
              - phone
              - adresse
            properties:
              firstName:
                type: string
              lastName:
                type: string
              phone:
                type: string
              adresse:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
            
  /user/{userId}:
    # binds a127 app logic to a route
    x-swagger-router-controller: userController
    
    get:
      description: get a user by Id
      # used as the method name of the controller
      operationId: getUserById
      parameters:
        - in: path
          name: userId
          type: string
          description: the id of the user
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
            
    put:
      description: Edit an existing User
      operationId: editUser
      parameters:
        - in: path
          name: userId
          type: string
          description: the id of the user
          required: true
        - in: body
          name: user
          description: The User to create
          schema:
            type: object
            required:
              - firstName
              - lastName
              - phone
              - adresse
            properties:
              firstName:
                type: string
              lastName:
                type: string
              phone:
                type: string
              adresse:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
    
  /reservation/{phoneId}:
    x-swagger-router-controller: reservationController
    post:
      description: Add a new Reservation
      # used as the method name of the controller
      operationId: addReservation
      parameters:
        - in: path
          name: phoneId
          type: string
          description: the id of the user who want a reservation
          required: true
        - in: body
          name: reservation
          description: The User to create
          schema:
            type: object
            required:
              - date
              - state
            properties:
              date:
                type: string
              state:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
    
  
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string